requirements.txt
python-telegram-bot==20.8
fastapi==0.110.0
uvicorn==0.29.0